thesis final
focuses on the use of binary instrumentation for real-time monitoring of IoT devices. The work explores the interplay between security and performance in IoT systems, assessing the impact and specific metrics of the overhead introduced by runtime monitoring techniques, particularly using Falco, a security tool for detecting intrusions and anomalies in systems. The research includes tests and benchmarks to quantify the overhead in various usage scenarios and provides a practical guide for integrating monitoring systems with the goal of enhancing the security and efficiency of IoT systems.